Single page templated items now can use contentful
Usecase: nav is global for all app pages ('single page templated' and not), generated with items pulled from a `contentful` object living in a template file as an `include`. Then the `contentful` object must be available in the single page template files via `locals` so that Spike could generate the nav inside.